Moved source files to "src" folder.
Added entrypoint script for running the game.
Patched module imports in source and test files.
