+++ test/test_settings.cfg
@@ -134,9 +134,17 @@ disease2 = Red
 disease3 = Yellow
 disease4 = Black
 
+[Log]
+enable_stream_log = true
+enable_file_log = true
+
 [Other]
 initial_city = London
 epidemics = 4
 rate = 2222334
 cubes = 30
-
+min_players = 2
+max_players = 4
+max_player_actions = 4
+outbreak_initial_level = 0
+outbreak_death_level = 8
